# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SessionLayer Control Plane API
  version: 0.1.0
  summary: Contract-first REST surface for the SessionLayer Control Plane.
  description: |
    The **SessionLayer Control Plane API** is the contract-first source of truth
    for the platform's REST surface (Design §13, FR-API-1). The UI is a *client* of
    this contract; server interfaces and typed clients are code-generated from this
    document, and CI fails if generated code drifts from the spec. Postgres is the
    single source of truth; the UI and the API share one write path (D11).

    **Complete contract.** This document is the complete resource surface: the
    config resources (`node-policies`, `rules`, `roles` + `role-bindings`, `cas`,
    `capability-defs`, `jit-policies`, `breakglass-policies`,
    `service-accounts`, `session-limit-policies`, `operator-settings`), the
    runtime resources (`nodes`, `gateways`, `sessions`, `session-leases`,
    `locks`, `jit-requests`, `join-tokens`, `otp`, device-flow, `pins`,
    `breakglass`, `recordings`, `audit-events`), and the operational probes.

    **Config vs runtime.** Config resources carry an `origin` provenance label
    (`api|ui|default`) and an optimistic-concurrency `version`; runtime resources
    (locks, sessions, issuance, audit) are never provenance-labelled. `locks` is an
    API-only runtime resource (Design §13).

    **Conventions.** URI versioning (`/v1`); cursor pagination on collections;
...
    compatibility policy.
...
servers:
...
security:
  - oidcBearer: []
  - clientCredentials: []
  - mtls: []
paths:
...
  /v1/locks:
    get:
      operationId: listLocks
      summary: List active locks (admin).
      description: |
        Lists the current unexpired locks. Platform-RBAC gated (`lock:read`) +
        audited (Design §8.3; FR-LOCK-1). A lock is the incident-response deny
        primitive; the deny reason is operator-only.
      tags:
        - locks
      responses:
        '200':
          description: The active locks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockList'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createLock
      summary: Create a lock (admin).
      description: |
        Creates an incident-response lock that blocks new session issuance AND
        tears down existing matching sessions on every Gateway (Design §8.3,
        §8.4; FR-LOCK-1/2), pushed over the actively-pushed deny-list. Ingest
        validation rejects an empty/unrecognised target; a fleet-wide lock
        requires an explicit `all: true`. Platform-RBAC gated (`lock:write`) +
        audited.
      tags:
        - locks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLockRequest'
      responses:
        '201':
          description: The lock was created and pushed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/locks/{lockId}:
    delete:
      operationId: releaseLock
      summary: Release a lock (admin).
      description: |
        Releases (deletes) a lock; the removal is pushed to every Gateway.
        Releasing a lock only stops future denial — it never resurrects a
        torn-down session. Platform-RBAC gated (`lock:write`) + audited.
      tags:
        - locks
      parameters:
        - name: lockId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The lock was released (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /v1/nodes:
    get:
      operationId: listNodes
      summary: List nodes (admin).
      description: |
        Lists enrolled nodes with their connectivity model, lifecycle status, and
        health (Design §12A; FR-NODE-1/2). Removed nodes are excluded by default.
        Platform-RBAC gated (`node:enroll`) + audited.
      tags:
        - nodes
      responses:
        '200':
          description: The enrolled nodes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeList'
        '403':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: registerNode
      summary: Register an agentless node (admin).
      description: |
        Enrolls an AGENTLESS node (Design §9.2/§9.3; FR-NODE-1): registers its dial
        address and host-identity material — a host-CA-signed host certificate OR
        an explicitly pinned host key (at least one; NEVER TOFU) — with no software
        install. An agent node instead joins via a join token
        (`POST /v1/join-tokens`). Enrollment MAY require approval (operator
        setting): when on, the node starts `pending` and is excluded from targeting
        until activated. Provisioning is a pure API flow (closes the dev-seed
        gap). Platform-RBAC gated (`node:enroll`) + audited.
      tags:
        - nodes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterNodeRequest'
      responses:
        '201':
          description: The node was registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/nodes/{nodeId}:
    get:
      operationId: getNode
      summary: Get a node (admin).
      description: Returns one node by id. Platform-RBAC gated (`node:enroll`) + audited.
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: removeNode
      summary: Remove a node (admin).
      description: |
        Deregisters the node (soft-remove: status `removed`, excluded from
        targeting; session/audit history preserved) and, for an AGENT node, REVOKES
        its agent credential — the identity is flipped off `active` and a covering
        Lock is pushed, so the generation counter + lock make a stale clone
        unusable and re-join cannot bypass the revocation (FR-NODE-3). Idempotent.
        Platform-RBAC gated (`node:remove`) + audited.
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The node was removed (idempotent).
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/nodes/{nodeId}/quarantine:
    post:
      operationId: quarantineNode
      summary: Quarantine a node (admin).
      description: |
        Immediately blocks NEW sessions to the node and tears down or drains
        EXISTING ones (FR-NODE-3), expressed as a top-tier Lock on the node
        (deny wins) actively pushed to every Gateway — fail-closed. The
        `existingSessions` policy selects `kill` (default, torn down at once) vs
        `drain` (finish, no new channels). Platform-RBAC gated (`node:quarantine`)
        + audited.
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuarantineNodeRequest'
      responses:
        '200':
          description: The node was quarantined and the lock pushed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: releaseQuarantine
      summary: Release a node from quarantine (admin).
      description: |
        Lifts a quarantine: releases the node Lock (pushed to every Gateway) and
        returns the node to `active`. Never resurrects a torn-down session.
        Idempotent. Platform-RBAC gated (`node:quarantine`) + audited.
      tags:
        - nodes
      parameters:
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The node was released from quarantine.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResource'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC bearer access via an ID/JWT token validated by the Control Plane
        (Design §5.7, FR-AUTH-17). The ID token is the authentication proof.
    clientCredentials:
      type: oauth2
      description: |
        OAuth 2.0 client-credentials for machine consumers (Design §5.6,
        FR-AUTH-12). `private_key_jwt`/mTLS client authentication is preferred
        over static secrets.
      flows:
        clientCredentials:
          tokenUrl: https://controlplane.sessionlayer.example/oauth2/token
          scopes: {}
    mtls:
      type: mutualTLS
      description: |
        Mutual TLS client-certificate authentication (Design §5.7, FR-AUTH-17),
        also used by the internal CP<->Gateway plane.
...
  responses:
    Problem:
      description: An error occurred. The body is an RFC 9457 problem document.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      title: RFC 9457 Problem Details
      description: |
        Problem details for HTTP APIs as defined by RFC 9457 (obsoletes RFC 7807).
        Error responses use media type `application/problem+json`.
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
          description: A URI reference identifying the problem type.
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: The HTTP status code.
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying the specific occurrence.
      additionalProperties: true
...
    LockMode:
      type: string
      title: Lock Mode
      description: |
        Lock enforcement mode (Design §8.3). `strict` tears down matching live
        sessions and blocks new ones; `best_effort` blocks new issuance but does
        not forcibly tear down an already-established session.
      enum:
        - strict
        - best_effort
    LockTarget:
      type: object
      title: Lock Target
      description: |
        What the lock matches (LockMatching facets; OR-matched). At least one
        facet must be non-empty, OR `all` must be true — a global lock is never
        implicit (ingest rejects an empty target).
      properties:
        identities:
          type: array
          items:
            type: string
          description: Resolved subject identities to deny.
        groups:
          type: array
          items:
            type: string
          description: SSO/OIDC groups to deny (membership match).
        nodeIds:
          type: array
          items:
            type: string
            format: uuid
          description: Target node ids to deny.
        principals:
          type: array
          items:
            type: string
          description: Linux logins (principals) to deny.
        nodeLabels:
          type: array
          items:
            type: string
          description: Node labels ("key=value") to deny.
        all:
          type: boolean
          description: An explicit fleet-wide lock (deny everything). Required for a target with no other facet.
      additionalProperties: false
    CreateLockRequest:
      type: object
      title: Create Lock Request
      required:
        - target
        - reason
      properties:
        target:
          $ref: '#/components/schemas/LockTarget'
        reason:
          type: string
          description: Operator/audit reason (mandatory). Never disclosed to the SSH user (§7.1).
        mode:
          $ref: '#/components/schemas/LockMode'
          description: Enforcement mode; omitted defaults to `strict` (backward-compatible).
        ttlSeconds:
          type: integer
          format: int64
          description: Optional lock lifetime in seconds; omitted or null = no expiry (indefinite until released).
      additionalProperties: false
    LockResource:
      type: object
      title: Lock
      required:
        - id
        - target
        - reason
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        target:
          $ref: '#/components/schemas/LockTarget'
        reason:
          type: string
        mode:
          $ref: '#/components/schemas/LockMode'
          description: Enforcement mode (`strict` or `best_effort`).
        expiresAt:
          type: string
          format: date-time
          description: Absolute expiry; absent when the lock has no TTL.
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: The admin identity that created the lock.
      additionalProperties: false
    LockList:
      type: object
      title: Lock List
      required:
        - locks
      properties:
        locks:
          type: array
          items:
            $ref: '#/components/schemas/LockResource'
      additionalProperties: false
...
    RegisterNodeRequest:
      type: object
      title: Register Node Request
      description: |
        Enrol an agentless node. `address` (host:port) is required; at least one of
        `hostCertificate` / `pinnedHostKey` is required (host-identity anchor — no
        TOFU, §9.3).
      required:
        - name
        - address
      properties:
        name:
          type: string
          description: The node's stable, unique human name (the enrollment key + the SSH addressing name).
        address:
          type: string
          description: The node sshd dial address, `host:port` (port defaults to 22 if omitted).
        labels:
          type: object
          additionalProperties:
            type: string
          description: Inventory labels ("key=value") used by data-plane rule/lock node selectors.
        hostCertificate:
          type: string
          description: The node's host-CA-signed OpenSSH host certificate line (the primary host-identity anchor).
        pinnedHostKey:
          type: string
          description: An explicitly pinned OpenSSH host public-key line (the fallback anchor when the node presents a plain host key).
        nodePolicyName:
          type: string
          description: Optional NodePolicy snapshot reference.
      additionalProperties: false
    QuarantineNodeRequest:
      type: object
      title: Quarantine Node Request
      required:
        - reason
      properties:
        reason:
          type: string
          description: Operator/audit reason (mandatory). Never disclosed to the SSH user (§7.1).
        existingSessions:
          type: string
          enum:
            - kill
            - drain
          default: kill
          description: How existing sessions are handled — `kill` (torn down at once) or `drain` (finish; no new channels).
        ttlSeconds:
          type: integer
          format: int64
          description: Optional quarantine-lock lifetime in seconds; omitted/null = indefinite until released.
      additionalProperties: false
    NodeResource:
      type: object
      title: Node
      required:
        - id
        - name
        - connectorKind
        - status
        - health
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        connectorKind:
          type: string
          enum:
            - agent
            - agentless
        status:
          type: string
          enum:
            - pending
            - active
            - quarantined
            - removed
        health:
          type: string
          enum:
            - unknown
            - healthy
            - unhealthy
            - unreachable
        address:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        owningGateway:
          type: string
          description: The Gateway currently owning the node's agent control channel (agent model); absent otherwise.
        statusReason:
          type: string
        statusChangedBy:
          type: string
        statusChangedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    NodeList:
      type: object
      title: Node List
      required:
        - nodes
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeResource'
      additionalProperties: false
...